# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Это API (swagger 2.0) системы автоматизации психолого-педагогическ\
    ого сопровождения субъектов образовательной деятельности для школы №18."
  version: "1.0.7"
  title: "PsychoSys18"
  contact:
    email: "antony.nor@yandex.ru"
host: "api-ps18.herokuapp.com"
basePath: "/"
...
schemes:
- "https"
paths:
...
  /addTest:
    post:
      tags:
      - "tests"
      summary: "Добавить оффлайн тест"
      operationId: "addTest"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Имя теста и файлы"
        required: true
        schema:
          $ref: "#/definitions/body_4"
      responses:
        200:
          description: "Статус запроса и ID добавленного теста"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_5_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /delTest:
    post:
      tags:
      - "tests"
      summary: "Удаление теста с файлами"
      operationId: "delTest"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID теста"
        required: true
        schema:
          $ref: "#/definitions/body_5"
      responses:
        200:
          description: "Статус запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /getTest:
    post:
      tags:
      - "tests"
      summary: "Просмотр списка тестирований"
      operationId: "getTest"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Статус запроса и информация о тестах"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_7_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /changeTestName:
    post:
      tags:
      - "tests"
      summary: "Изменение названия теста"
      operationId: "changeTestName"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID теста и его новое имя"
        required: true
        schema:
          $ref: "#/definitions/body_6"
      responses:
        200:
          description: "Статус запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /accessTest:
    post:
      tags:
      - "tests"
      summary: "Открытие/Закрытие доступа к тестам ученикам"
      operationId: "accessTest"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID теста, ID студента и статус открытия доступа"
        required: true
        schema:
          $ref: "#/definitions/body_7"
      responses:
        200:
          description: "Статус запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /addTestResult:
    post:
      tags:
      - "tests"
      summary: "Добавление результатов тестирований"
      operationId: "addTestResult"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID теста, ID студента и файлы с результатами"
        required: true
        schema:
          $ref: "#/definitions/body_8"
      responses:
        200:
          description: "Статус запроса и ID добавленного результата"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_8_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /getTestResult:
    post:
      tags:
      - "tests"
      summary: "Поиск результатов тестирований по по тестам и ученикам"
      operationId: "getTestResult"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID теста и/или ID студента"
        required: true
        schema:
          $ref: "#/definitions/body_9"
      responses:
        200:
          description: "Статус запроса и ID найденных результатов тестирований"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_8_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
  /delTestResult:
    post:
      tags:
      - "tests"
      summary: "Удаление результатов тестирования вместе с фалами"
      operationId: "delTestResult"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID результата тестирования"
        required: true
        schema:
          $ref: "#/definitions/body_10"
      responses:
        200:
          description: "Статус запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Tests"
    x-swagger-router-controller: "tests"
...
  /getTeachers:
    post:
      tags:
      - "users"
      summary: "Получение списка учителей для назначения классного руководства (Admin)"
      operationId: "getTeachers"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Статус запроса и массив преподавателей"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_21_payload"
        401:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getPersonalData:
    post:
      tags:
      - "users"
      summary: "Получение персональных данных (ALL)"
      operationId: "getPersonalData"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Данные пользователя."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_22_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getPersonsToBeRec:
    post:
      tags:
      - "users"
      summary: "Получение списка сотрудников доступных для записи (ALL)"
      operationId: "getPersonsToBeRec"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Данные пользователя."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_23_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getUser:
    post:
      tags:
      - "users"
      summary: "Получение данных пользователя"
      operationId: "getUser"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID пользователя"
        required: true
        schema:
          $ref: "#/definitions/body_25"
      responses:
        200:
          description: "Данные пользователя."
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                $ref: "#/definitions/inline_response_200_24_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getStudent:
    post:
      tags:
      - "users"
      summary: "Поиск учеников с фильтрацией"
      operationId: "getStudent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "параметры поиска ученика"
        required: true
        schema:
          $ref: "#/definitions/body_26"
      responses:
        200:
          description: "Список учеников, удовлетворяющих критериям поиска"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_22_payload_prnt_data_prnt_childrens"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getEmployee:
    post:
      tags:
      - "users"
      summary: "Поиск сотрудников с фильтрацией"
      operationId: "getEmployee"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Параметры поиска сотрудника"
        required: true
        schema:
          $ref: "#/definitions/body_27"
      responses:
        200:
          description: "Список сотрудников, удовлетворяющих критериям поиска"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_26_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /getParent:
    post:
      tags:
      - "users"
      summary: "Поиск родителей с фильтрацией"
      operationId: "getParent"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Параметры поиска родителя"
        required: true
        schema:
          $ref: "#/definitions/body_28"
      responses:
        200:
          description: "Список родителей, удовлетворяющих критериям поиска"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
              payload:
                type: "array"
                items:
                  $ref: "#/definitions/inline_response_200_27_payload"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
  /bindParentAndStud:
    post:
      tags:
      - "users"
      summary: "Привязка ученика к родителю"
      operationId: "bindParentAndStud"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "ID ученика и родителя"
        required: true
        schema:
          $ref: "#/definitions/body_29"
      responses:
        200:
          description: "Статус выполнения запроса"
          schema:
            type: "object"
            required:
            - "status"
            properties:
              status:
                type: "string"
                example: "OK"
        400:
          description: "Invalid data"
        default:
          description: "Unknown Error"
      security:
      - JWT: []
      x-swagger-router-controller: "Users"
    x-swagger-router-controller: "users"
...
securityDefinitions:
  Basic:
    type: "basic"
  JWT:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Role:
    type: "array"
    items:
      type: "string"
      example: ""
      description: "Value:\n * Admin\n * Student\n * Parent\n * Teacher\n * Psychologist\n"
...
  addTest_file_arr:
    required:
    - "file"
    - "file_name"
    properties:
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file:
        type: "string"
        format: "base64"
  body_4:
    type: "object"
    required:
    - "file_arr"
    - "tst_name"
    properties:
      tst_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_arr:
        type: "array"
        items:
          $ref: "#/definitions/addTest_file_arr"
  inline_response_200_5_payload:
    required:
    - "tst_id"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
    example:
      tst_id: 1
...
  body_5:
    type: "object"
    required:
    - "tst_id"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
...
  inline_response_200_7_payload:
    required:
    - "tst_id"
    - "tst_name"
    - "tst_online"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      tst_name:
        type: "string"
        example: "Тест по профорейнтации"
      tst_online:
        type: "boolean"
        example: false
    example:
      tst_name: "Тест по профорейнтации"
      tst_online: false
      tst_id: 1
...
  body_6:
    type: "object"
    required:
    - "tst_id"
    - "tst_name"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      tst_name:
        type: "string"
        example: "Новый Тест по профорейнтации"
  body_7:
    type: "object"
    required:
    - "std_id"
    - "tst_id"
    - "tst_open_access"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      tst_open_access:
        type: "boolean"
        example: true
  addTestResult_file_arr:
    required:
    - "file"
    - "file_name"
    - "file_path"
    properties:
      file_name:
        type: "string"
        example: "Тест по профорейнтации"
      file_path:
        type: "string"
        example: "/folder1"
      file:
        type: "string"
        format: "base64"
  body_8:
    type: "object"
    required:
    - "file_arr"
    - "std_id"
    - "tst_id"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      file_arr:
        type: "array"
        items:
          $ref: "#/definitions/addTestResult_file_arr"
  inline_response_200_8_payload:
    required:
    - "tr_id"
    properties:
      tr_id:
        type: "integer"
        format: "int32"
        example: 1
    example:
      tr_id: 1
...
  body_9:
    type: "object"
    properties:
      tst_id:
        type: "integer"
        format: "int32"
        example: 1
      std_id:
        type: "integer"
        format: "int32"
        example: 3
...
  body_10:
    type: "object"
    required:
    - "tr_id"
    properties:
      tr_id:
        type: "integer"
        format: "int32"
        example: 1
...
  inline_response_200_21_payload:
    required:
    - "emp_data"
    - "pepl_data"
    - "pepl_id"
    - "pepl_login"
    - "role_array"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
      pepl_login:
        type: "string"
        example: "Admin"
      role_array:
        $ref: "#/definitions/Role"
      pepl_data:
        $ref: "#/definitions/People"
      emp_data:
        $ref: "#/definitions/Employee"
      pst_arr:
        type: "array"
        items:
          type: "string"
          example: ""
      class_arr:
        type: "array"
        items:
          type: "string"
          example: ""
    example:
      pepl_login: "Admin"
      pst_arr:
      - ""
      - ""
      emp_data:
        emp_vk: "empadmin"
        emp_skype: "emplAdmin"
        emp_viber: "9568734554"
        emp_hangouts: "admin@gmail.com"
        emp_date_enrollment: "1997-11-22"
        emp_discord: "emplAdmin#1232"
      pepl_id: 1
      class_arr:
      - ""
      - ""
      role_array: ""
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
...
  inline_response_200_22_payload_prnt_data_prnt_childrens:
    required:
    - "pepl_data"
    - "pepl_id"
    - "pepl_login"
    - "std_data"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 3
      pepl_login:
        type: "string"
        example: "Student"
      pepl_data:
        $ref: "#/definitions/People"
      std_data:
        $ref: "#/definitions/Student"
    example:
      pepl_login: "Student"
      pepl_id: 3
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
      std_data:
        std_stayed_two_year: 0
        std_class: "7Б"
        std_date_issue: "2023-06-01"
...
  inline_response_200_22_payload:
    required:
    - "pepl_data"
    - "pepl_login"
    - "role_array"
    properties:
      pepl_login:
        type: "string"
        example: "Admin"
      role_array:
        $ref: "#/definitions/Role"
      pepl_data:
        $ref: "#/definitions/People"
      emp_data:
        $ref: "#/definitions/inline_response_200_22_payload_emp_data"
      prnt_data:
        $ref: "#/definitions/inline_response_200_22_payload_prnt_data"
      std_data:
        $ref: "#/definitions/Student"
    description: "Данные пользователя"
    example:
      pepl_login: "Admin"
      emp_data:
        main_data:
          emp_vk: "empadmin"
          emp_skype: "emplAdmin"
          emp_viber: "9568734554"
          emp_hangouts: "admin@gmail.com"
          emp_date_enrollment: "1997-11-22"
          emp_discord: "emplAdmin#1232"
        pst_arr:
        - ""
        - ""
      role_array: ""
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
      prnt_data:
        main_data:
          prnt_flat: "45"
          prnt_home: "17Б"
          prnt_city: "Липецк"
          prnt_street: "Ангарская"
        prnt_childrens:
        - pepl_login: "Student"
          pepl_id: 3
          pepl_data:
            pepl_first_name: "Иван"
            pepl_gender: true
            pepl_second_name: "Иванов"
            pepl_last_name: "Иванович"
            pepl_email: "admin@mail.ru"
            pepl_phone: "9568734554"
            pepl_birthday: "1977-01-19"
          std_data:
            std_stayed_two_year: 0
            std_class: "7Б"
            std_date_issue: "2023-06-01"
        - pepl_login: "Student"
          pepl_id: 3
          pepl_data:
            pepl_first_name: "Иван"
            pepl_gender: true
            pepl_second_name: "Иванов"
            pepl_last_name: "Иванович"
            pepl_email: "admin@mail.ru"
            pepl_phone: "9568734554"
            pepl_birthday: "1977-01-19"
          std_data:
            std_stayed_two_year: 0
            std_class: "7Б"
            std_date_issue: "2023-06-01"
      std_data:
        std_stayed_two_year: 0
        std_class: "7Б"
        std_date_issue: "2023-06-01"
...
  inline_response_200_23_payload:
    required:
    - "pepl_first_name"
    - "pepl_id"
    - "pepl_last_name"
    - "pepl_login"
    - "pepl_second_name"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
      pepl_login:
        type: "string"
        example: "Admin"
      pepl_second_name:
        type: "string"
        example: "Иванов"
      pepl_first_name:
        type: "string"
        example: "Иван"
      pepl_last_name:
        type: "string"
        example: "Иванович"
      pst_arr:
        type: "array"
        items:
          type: "string"
          example: ""
    description: "Данные пользователя"
    example:
      pepl_first_name: "Иван"
      pepl_login: "Admin"
      pepl_second_name: "Иванов"
      pst_arr:
      - ""
      - ""
      pepl_id: 1
      pepl_last_name: "Иванович"
...
  body_25:
    type: "object"
    required:
    - "pepl_id"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
  inline_response_200_24_payload:
    required:
    - "pepl_data"
    - "role_array"
    properties:
      role_array:
        $ref: "#/definitions/Role"
      pepl_data:
        $ref: "#/definitions/People"
      emp_data:
        $ref: "#/definitions/Employee"
      prnt_data:
        $ref: "#/definitions/Parent"
      std_data:
        $ref: "#/definitions/Student"
    description: "Данные пользователя"
    example:
      emp_data:
        emp_vk: "empadmin"
        emp_skype: "emplAdmin"
        emp_viber: "9568734554"
        emp_hangouts: "admin@gmail.com"
        emp_date_enrollment: "1997-11-22"
        emp_discord: "emplAdmin#1232"
      role_array: ""
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
      prnt_data:
        prnt_flat: "45"
        prnt_home: "17Б"
        prnt_city: "Липецк"
        prnt_street: "Ангарская"
      std_data:
        std_stayed_two_year: 0
        std_class: "7Б"
        std_date_issue: "2023-06-01"
...
  body_26:
    type: "object"
    properties:
      emp_id:
        type: "integer"
        format: "int32"
        example: 1
      std_parallel:
        type: "integer"
        format: "int32"
        example: 7
      std_letter:
        type: "string"
        example: "Б"
      prnt_id:
        type: "integer"
        format: "int32"
        example: 2
      pepl_login:
        type: "string"
        example: "Student"
      pepl_second_name:
        type: "string"
        example: "Тарасов"
      pepl_first_name:
        type: "string"
        example: "Семен"
      pepl_last_name:
        type: "string"
        example: "Львович"
      pepl_gender:
        type: "boolean"
        example: true
      pepl_birthday:
        type: "string"
        format: "date"
        example: "2005-05-12"
      pepl_phone:
        type: "string"
        example: "6756776565"
      pepl_email:
        type: "string"
        example: "stud@mail.ru"
...
  getEmployee_post_array:
    properties:
      pst_id:
        type: "integer"
        format: "int32"
        example: 1
  body_27:
    type: "object"
    properties:
      role_array:
        $ref: "#/definitions/Role"
      post_array:
        type: "array"
        items:
          $ref: "#/definitions/getEmployee_post_array"
      pepl_login:
        type: "string"
        example: "Admin"
      pepl_second_name:
        type: "string"
        example: "Иванов"
      pepl_first_name:
        type: "string"
        example: "Иван"
      pepl_last_name:
        type: "string"
        example: "Иванович"
      pepl_gender:
        type: "boolean"
        example: true
      pepl_birthday:
        type: "string"
        format: "date"
        example: "1977-01-19"
      pepl_phone:
        type: "string"
        example: "9568734554"
      pepl_email:
        type: "string"
        example: "admin@mail.ru"
  inline_response_200_26_payload:
    required:
    - "emp_data"
    - "pepl_data"
    - "pepl_id"
    - "pepl_login"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 1
      pepl_login:
        type: "string"
        example: "Admin"
      pepl_data:
        $ref: "#/definitions/People"
      emp_data:
        $ref: "#/definitions/Employee"
    example:
      pepl_login: "Admin"
      emp_data:
        emp_vk: "empadmin"
        emp_skype: "emplAdmin"
        emp_viber: "9568734554"
        emp_hangouts: "admin@gmail.com"
        emp_date_enrollment: "1997-11-22"
        emp_discord: "emplAdmin#1232"
      pepl_id: 1
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
...
  body_28:
    type: "object"
    properties:
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      prnt_confirm:
        type: "integer"
        format: "int32"
        example: 0
        description: "Value:\n * 0 - на рассмотрении\n * 1 - подтверждено\n * 2 -\
          \ отклонено\n"
      pepl_login:
        type: "string"
        example: "Parent"
      pepl_second_name:
        type: "string"
        example: "Смирнова"
      pepl_first_name:
        type: "string"
        example: "Ольга"
      pepl_last_name:
        type: "string"
        example: "Викторовна"
      pepl_gender:
        type: "boolean"
        example: false
      pepl_birthday:
        type: "string"
        format: "date"
        example: "1990-02-20"
      pepl_phone:
        type: "string"
        example: "5345345543"
      pepl_email:
        type: "string"
        example: "parent@mail.ru"
  inline_response_200_27_payload:
    required:
    - "pepl_data"
    - "pepl_id"
    - "pepl_login"
    - "prnt_data"
    properties:
      pepl_id:
        type: "integer"
        format: "int32"
        example: 2
      pepl_login:
        type: "string"
        example: "Parent"
      pepl_data:
        $ref: "#/definitions/People"
      prnt_data:
        $ref: "#/definitions/Parent"
    example:
      pepl_login: "Parent"
      pepl_id: 2
      pepl_data:
        pepl_first_name: "Иван"
        pepl_gender: true
        pepl_second_name: "Иванов"
        pepl_last_name: "Иванович"
        pepl_email: "admin@mail.ru"
        pepl_phone: "9568734554"
        pepl_birthday: "1977-01-19"
      prnt_data:
        prnt_flat: "45"
        prnt_home: "17Б"
        prnt_city: "Липецк"
        prnt_street: "Ангарская"
...
  body_29:
    type: "object"
    required:
    - "prnt_id"
    - "std_id"
    properties:
      std_id:
        type: "integer"
        format: "int32"
        example: 3
      prnt_id:
        type: "integer"
        format: "int32"
        example: 2
...